--- src/pages/UserList.jsx
@@ -1,29 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import axios from 'axios';
-
-export default function UserList() {
-  const [users, setUsers] = useState([]);
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    axios.get(`${process.env.REACT_APP_AUTH_URL}/api/users`)
-      .then(response => {
-        setUsers(response.data);
-      })
-      .catch(error => console.error('Error fetching users:', error));
-  }, []);
-
-  return (
-    <div>
-      <h1>User List</h1>
-      <ul>
-        {users.map(user => (
-          <li key={user.memberId} onClick={() => navigate(`/api/users/${user.memberId}/posts`)}>
-            {user.nickname}
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
-}
