+++ src/pages/posts/Pagination.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { Button } from "../components/button";
+import { Button } from "../../components/button";
 
 export default function Pagination({ page, setPage, data }) {
   const postsPerPage = 10;
